requirements.txt
# Core ML libraries
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
diffusers>=0.21.0
accelerate>=0.24.0
peft>=0.6.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
scikit-image>=0.21.0

# Face recognition and quality control
face-recognition>=1.3.0
insightface>=0.7.3
onnxruntime-gpu>=1.16.0

# Data handling
pandas>=2.0.0
pyyaml>=6.0
tqdm>=4.66.0

# Utilities
psutil>=5.9.0
pynvml>=11.5.0
python-dotenv>=1.0.0

# Optional optimizations
xformers>=0.0.22; sys_platform != "win32"
# flash-attn; sys_platform != "win32"  # Requires special build

# Logging and monitoring
wandb>=0.15.0  # Optional, for training monitoring
tensorboard>=2.14.0

# EXIF handling
exifread>=3.0.0
piexif>=1.1.3

# ControlNet and adapters (will be installed separately if needed)
# controlnet-aux
# ip-adapter

